+++ R/group_var.R
@@ -11,17 +11,19 @@
 #'          or \code{\link{rec_pattern}} and \code{\link{rec}} for another
 #'          convenbient way of recoding variables into smaller groups.
 #'
-#' @param groupsize Numeric; group-size, i.e. the range for grouping. By default,
-#'          for each 5 categories of \code{x} a new group is defined, i.e. \code{groupsize=5}.
-#'          Use \code{groupsize = "auto"} to automatically resize a variable into
+#' @param size Numeric; group-size, i.e. the range for grouping. By default,
+#'          for each 5 categories of \code{x} a new group is defined, i.e. \code{size = 5}.
+#'          Use \code{size = "auto"} to automatically resize a variable into
 #'          a maximum of 30 groups (which is the ggplot-default grouping when
-#'          plotting histograms). Use \code{groupcount} to determine the amount
+#'          plotting histograms). Use \code{n} to determine the amount
 #'          of groups.
 #' @param right.interval Logical; if \code{TRUE}, grouping starts with the lower
-#'          bound of \code{groupsize}. See 'Details'.
-#' @param groupcount Sets the maximum number of groups that are defined when auto-grouping is on
-#'          (\code{groupsize="auto"}). Default is 30. If \code{groupsize} is not set to \code{"auto"},
+#'          bound of \code{size}. See 'Details'.
+#' @param n Sets the maximum number of groups that are defined when auto-grouping is on
+#'          (\code{size = "auto"}). Default is 30. If \code{size} is not set to \code{"auto"},
 #'          this argument will be ignored.
+#' @param groupsize Deprecated. Use \code{size} instead.
+#' @param groupcount Deprecated. Use \code{n} instead.
 #'
 #' @inheritParams to_factor
 #' @inheritParams rec
@@ -32,26 +34,26 @@
 #'         }
 #'
 #' @note Variable label attributes (see, for instance, \code{\link[sjlabelled]{set_label}}) are preserved.
-#'       Usually you should use the same values for \code{groupsize} and
+#'       Usually you should use the same values for \code{size} and
 #'       \code{right.interval} in \code{group_label()} as used in the
 #'       \code{group_var} function if you want matching labels for the related
 #'       recoded variable.
 #'
-#' @details If \code{groupsize} is set to a specific value, the variable is recoded
-#'            into several groups, where each group has a maximum range of \code{groupsize}.
+#' @details If \code{size} is set to a specific value, the variable is recoded
+#'            into several groups, where each group has a maximum range of \code{size}.
 #'            Hence, the amount of groups differ depending on the range of \code{x}.
 #'            \cr \cr
-#'            If \code{groupsize = "auto"}, the variable is recoded into a maximum of
-#'            \code{groupcount} groups. Hence, independent from the range of
+#'            If \code{size = "auto"}, the variable is recoded into a maximum of
+#'            \code{n} groups. Hence, independent from the range of
 #'            \code{x}, always the same amount of groups are created, so the range
 #'            within each group differs (depending on \code{x}'s range).
 #'            \cr \cr
 #'            \code{right.interval} determins which boundary values to include when
 #'            grouping is done. If \code{TRUE}, grouping starts with the \strong{lower
-#'            bound} of \code{groupsize}. For example, having a variable ranging from
+#'            bound} of \code{size}. For example, having a variable ranging from
 #'            50 to 80, groups cover the ranges from  50-54, 55-59, 60-64 etc.
 #'            If \code{FALSE} (default), grouping starts with the \code{upper bound}
-#'            of \code{groupsize}. In this case, groups cover the ranges from
+#'            of \code{size}. In this case, groups cover the ranges from
 #'            46-50, 51-55, 56-60, 61-65 etc. \strong{Note:} This will cover
 #'            a range from 46-50 as first group, even if values from 46 to 49
 #'            are not present. See 'Examples'.
@@ -62,11 +64,11 @@
 #'
 #' @examples
 #' age <- abs(round(rnorm(100, 65, 20)))
-#' age.grp <- group_var(age, groupsize = 10)
+#' age.grp <- group_var(age, size = 10)
 #' hist(age)
 #' hist(age.grp)
 #'
-#' age.grpvar <- group_labels(age, groupsize = 10)
+#' age.grpvar <- group_labels(age, size = 10)
 #' table(age.grp)
 #' print(age.grpvar)
 #'
@@ -86,7 +88,7 @@
 #' library(dplyr)
 #' efc %>%
 #'   select(e17age, c12hour, c160age) %>%
-#'   group_var(groupsize = 20)
+#'   group_var(size = 20)
 #'
 #' # create vector with values from 50 to 80
 #' dummy <- round(runif(200, 50, 80))
@@ -97,8 +99,21 @@
 #'
 #' @importFrom purrr map
 #' @export
-group_var <- function(x, ..., groupsize = 5, as.num = TRUE, right.interval = FALSE,
-                      groupcount = 30, append = FALSE, suffix = "_gr") {
+group_var <- function(x, ..., size = 5, as.num = TRUE, right.interval = FALSE,
+                      n = 30, append = FALSE, suffix = "_gr", groupsize, groupcount) {
+
+  # check deprecated arguments
+  if (!missing(groupcount)) {
+    message("Argument `groupcount` is deprecated. Please use `n` instead.")
+    n <- groupcount
+  }
+
+  if (!missing(groupsize)) {
+    message("Argument `groupsize` is deprecated. Please use `size` instead.")
+    size <- groupsize
+  }
+
+
   # evaluate arguments, generate data
   .dat <- get_dot_data(x, dplyr::quos(...))
 
@@ -111,10 +126,10 @@ group_var <- function(x, ..., groupsize = 5, as.num = TRUE, right.interval = FAL
     for (i in colnames(.dat)) {
       x[[i]] <- g_v_helper(
         x = .dat[[i]],
-        groupsize = groupsize,
+        groupsize = size,
         as.num = as.num,
         right.interval = right.interval,
-        groupcount = groupcount
+        groupcount = n
       )
     }
 
@@ -131,10 +146,10 @@ group_var <- function(x, ..., groupsize = 5, as.num = TRUE, right.interval = FAL
   } else {
     x <- g_v_helper(
       x = .dat,
-      groupsize = groupsize,
+      groupsize = size,
       as.num = as.num,
       right.interval = right.interval,
-      groupcount = groupcount
+      groupcount = n
     )
   }
 
@@ -164,7 +179,18 @@ g_v_helper <- function(x, groupsize, as.num, right.interval, groupcount) {
 
 #' @rdname group_var
 #' @export
-group_labels <- function(x, ..., groupsize = 5, right.interval = FALSE, groupcount = 30) {
+group_labels <- function(x, ..., size = 5, right.interval = FALSE, n = 30, groupsize, groupcount) {
+  # check deprecated arguments
+  if (!missing(groupcount)) {
+    message("Argument `groupcount` is deprecated. Please use `n` instead.")
+    n <- groupcount
+  }
+
+  if (!missing(groupsize)) {
+    message("Argument `groupsize` is deprecated. Please use `size` instead.")
+    size <- groupsize
+  }
+
   # evaluate arguments, generate data
   .dat <- get_dot_data(x, dplyr::quos(...))
 
@@ -173,17 +199,17 @@ group_labels <- function(x, ..., groupsize = 5, right.interval = FALSE, groupcou
     return(
       purrr::map(.dat, ~ g_l_helper(
         x = .x,
-        groupsize = groupsize,
+        groupsize = size,
         right.interval = right.interval,
-        groupcount = groupcount
+        groupcount = n
     )))
 
   } else {
     x <- g_l_helper(
       x = .dat,
-      groupsize = groupsize,
+      groupsize = size,
       right.interval = right.interval,
-      groupcount = groupcount
+      groupcount = n
     )
   }